chore: Update Facebook SDK initialization in Facebook.jsx
Improved the initialization of the Facebook SDK in the Facebook.jsx component. Replaced the previous code with a new script that loads the SDK asynchronously and includes the necessary environment variables. This change ensures correct initialization of the Facebook SDK and better compatibility with the updated Facebook API version 'v8.0'.